refactor(naming): one set of naming rules, used by prepare as well
`validate_object_naming` exported nothing but its MCP wrapper, so the 13 rules
inside it were unreachable. prepare could not call them and grew TWO hand-rolled
validators instead — and they disagreed with it. Live, on the same input:

  prepare(mode="create", objectName="CustTable", objectType="table")
    -> "✅ Naming looks valid."
  validate_object_naming(proposedName="CustTable", objectType="table")
    -> "⚠ Proposed name does not start with model prefix "ConChain""

and worse, for the case that matters most:

  prepare(mode="change", proposedName="CustTable_Ext")
    -> "ℹ️  Confirm naming follows your convention"   — no check at all
  validate_object_naming(…, "table-extension", base "CustTable")
    -> ERROR + the expected name "CustTable.ConChainExtension"

The rules now live in utils/objectNamingRules.ts and all three callers run them.
NOTHING was deleted and nothing was rewritten: the rule body was MOVED by slicing
the file, so the extension-shape rules, the #833 prefix-conflict warning, the
#892/#901 model-token handling, the #686 case-insensitive probes and the
extension-type reinterpretation are the same code they were. The 97 tests that
encode those lessons pass unchanged — that is the proof of the move.

What deliberately stayed local:

- prepareCreate keeps its own final-name guard. The shared rules see the name the
  caller TYPED; that guard covers the name the prefix step COMPOSED, which is
  where an unrepresentable character enters a name nobody typed (#892/#901).
- prepareCreate keeps its own collision check, which is better than the
  validator's — it probes the base name AND the final one.
- prepare renders errors, warnings and ONE suggestion rather than the validator's
  full block: its responses already hit the 5 000-char cap and were truncating.

prepareChange also loses a rule that was actively wrong: `!proposedName
.includes(modelName)` against the RAW model name is the #892/#901 class of bug,
fixed in the shared rules long ago.

New tests/tools/prepareNamingAgreement.test.ts pins prepare to the validator.
prefixResolutionAgreement.test.ts pins workspace_info to it; prepare was the pair
nothing covered, which is how they drifted.

Verified live against the real 2.5 GB index: both paths now answer a malformed
extension name with the same rule and the same expected name.

330 test files, 4822 passed.
